EWM: fix

+++ ewm-core/src/main/java/ru/practicum/ewmcore/model/participationRequest/ParticipationRequestDto.java
@@ -7,6 +7,6 @@ public class ParticipationRequestDto {
     private Long id;
     private String created; //дата создания для дто (в формате "yyyy-MM-dd HH:mm:ss")
     private Long event;
-    private Integer requester;
+    private Long requester;
     private ParticipationRequestStateEnum status;
 }
